gcc: Upgrade version from 4.7.2 to 4.8.1.
QorIQ SDK v1.6 will use GCC 4.8.1. The current version on the master
branch is 4.7.2.

For GCC 4.8.1, the Graphite framework for loop optimizations uses isl,
instead of PPL. To enable the framework, we need CLooG version 0.18.0
and isl version 0.11.1.

This patch does the following:

1. Adds recipes for GCC 4.8;
2. Adds recipes for isl and CLooG;
3. Deletes recipes for GCC 4.7;
4. Deletes recipes for PPL and CLooG-PPL.

In addition, it updates the SHA-1 hash for EGLIBC to HEAD, in order to
incorporate a commit needed for GCC 4.8.1 integration.

Signed-off-by: Cosmin Paraschiv <cosmin.paraschiv@freescale.com>
Change-Id: I0e835b73912fe8d6a6e9a12b43de2d330afef651
Reviewed-on: http://git.am.freescale.net:8181/7061
Reviewed-by: Chunrong Guo <B40290@freescale.com>
Reviewed-by: Zhenhua Luo <zhenhua.luo@freescale.com>
Tested-by: Zhenhua Luo <zhenhua.luo@freescale.com>