gdb-cross: Fix incorrect exec-prefix Python path.
In the configure stage, GDB defines WITH_PYTHON_PATH with the value of the
'--exec-prefix' case statement. The macro is used later, during compilation,
to compute the path for Python run-time libraries and the path for the Python
executable.

If the sysroot is relocated, GDB has the ability to recompute those paths,
in order to have Python support, but it relies on at least one of two things:

1) the user must set the PYTHONHOME variable;
2) the aforementioned paths computed from WITH_PYTHON_PATH must be correct.

If neither is true, GDB can't recompute the Python paths.

The relocation scenario is most plausible for the SDK, which packages
gdb-cross-canadian, not gdb-cross, but the patch doesn't do any harm,
if we want to be politically correct.

Signed-off-by: Cosmin Paraschiv <cosmin.paraschiv@freescale.com>
Change-Id: Iad38c8db68da0897b3751c79272408d8a9c81db2
Reviewed-on: http://git.am.freescale.net:8181/8966
Reviewed-by: Zhenhua Luo <zhenhua.luo@freescale.com>
Tested-by: Zhenhua Luo <zhenhua.luo@freescale.com>