gcc-cross-canadian: Add '-ffreestanding' to BUILD_CPPFLAGS.
Ubuntu 13.10 comes with GCC 4.8.1. The 4.8 release series brought some changes
regarding the preprocessor. It may now pre-include a file that defines certain
macros for the entirety of the translation unit. On Linux, 'stdc-predef.h' is
pre-included.

This is a problem for Yocto, since GCC ends up mixing include files from
the host's C standard library (EGLIBC 2.17) and from Yocto (EGLIBC 2.15).

The pre-include can be disabled with the use of '-ffreestanding'. It asserts
that compilation takes place in a freestanding environment, in which libc may
not exist. Consequently, it will only use include files from Yocto's sysroot.

Signed-off-by: Cosmin Paraschiv <cosmin.paraschiv@freescale.com>
Change-Id: I0c5e033fa91093194f920ea00b4a57e19bc81382
Reviewed-on: http://git.am.freescale.net:8181/9107
Reviewed-by: Ting Liu <ting.liu@freescale.com>
Tested-by: Ting Liu <ting.liu@freescale.com>