binutils: lower priority for multilib symlinks
Give lower priority for binutils update-alternatives when
building a multitilib variant of the recipe because the
default symlink should corespond with the binary compiled
with the default tune for the target architecture.
This configuration is skipped if the package is nativesdk or
is not multilib.

Eg: /usr/bin/strip should always point to /usr/bin/powerpc-fsl-linux-strip
on 32 bit architectures even if we install the 64 bit variant of binutils.

Signed-off-by: Alexandru-Cezar Sardan <alexandru.sardan@freescale.com>
Change-Id: Ib333b3df5434fb08441d454533c59cb931a667cb
Reviewed-on: http://git.am.freescale.net:8181/12226
Reviewed-by: Ting Liu <ting.liu@freescale.com>
Tested-by: Ting Liu <ting.liu@freescale.com>