Fix Autoconf version mismatch error for GCC and Binutils
Binutils requires an exact version because files generated with autoconf are
commited in the upstream repository and there would be differences in the
generated files between versions.
This ensures that everyone is using the same version of autoconf.

This patch rewritesthe  _GCC_AUTOCONF_VERSION variable so that it corresponds
with the version used in Yocto.

Signed-off-by: Alexandru-Cezar Sardan <alexandru.sardan@freescale.com>
Change-Id: I3333233b48a943b5ffbdfdb63e287b4648961140
Reviewed-on: http://git.am.freescale.net:8181/12953
Reviewed-by: Zhenhua Luo <zhenhua.luo@freescale.com>
Tested-by: Zhenhua Luo <zhenhua.luo@freescale.com>